Add `nested-in-python` representation
It is hard to deal with a single regexp in code. For example, the
regular expression for RFC 3987 is >5K characters long.

We introduce an additional representation where the regular expression
is compiled as nested strings in Python (and potentially other languages
in the future) to improve the readability of the resulting generated
code.