Use Tarjan's depth-first topological sort
This patch changes the algorithm used for constructing and traversing
the dependency graph of the rules. The original implementation did not
report the cycles.

We use now Tarjan's algorithm based on depth-first for the dependency
graph.

Additionally, the rules from RFC 5234 are now properly short-circuited.
This obscured a bug related to the cycles in the ABNF.