Replace repetition range with shorter alternatives (#16)
This patch improves the readability of the resulting expressions by
shortening repetition ranges where possible.

The range is replaced with the exact count if the minimum and maximum
occurence coincide. If there is no maximum range, but only the minimum
range of 1 is given, the repetition is defined with a `+`.