Upgrade to abnf 2.0.0
We upgrade to the latest version of abnf to keep up-to-date with the
dependencies.

During the upgrade, we had to re-structure the code a bit and discovered
a bug where the letter ranges were erroneously inferred. Namely, we
missed the cases where the range start and end do not involve letters,
but overlap with one or more letter ranges. This patch includes the fix
of the bug as well as the re-structuring.

We also had to drop support for Python 3.6 as the module [abnf] also
dropped support for it in version 2.0.0.

[abnf]: https://github.com/declaresub/abnf