improve regexp generation for `#` (#50)
The current code escaped `#` in the regular expressions. However, this
conflicts with other regex engines, such as unicode-aware JavaScript's
one. We remove the escaping for `#` in this change to make the generated
patterns usable on more platforms.

In addition, this escaping behavior is harmonized between the nested
python and single regexp generators.

Fixes #49.
Fixes #46.